Move CAS1 seed test application creation logic
Move logic into service in preparation for reuse
